requirements.txt
huggingface_hub
streamlit
google-adk
openai
pymupdf4llm